backend/requirements.txt
fastapi~=0.109.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
openai==1.12.0
python-multipart==0.0.9
pydantic>=2.4.1,<2.6
aiogram==3.3.0
python-telegram-bot>=20.0